v1.3 (Pre-release)
Fixed a small bug with the UV map due to changes in the tail, as well as Facial expression (formerly called Gesture_mimics_off). Made manual control of avatar's eyes, extensive tail control (three modes of operation and with manual disabling of colliders for floor and legs), tweaked weights for jeans, pants and shorts, and especially neck and arms of avatar's body, added additional bones to ears (now ears are more elastic and pleasant to move >_<), Added possibility to disable “Pet expressions” (disabling “shyness” and nose pressing sound). Due to inattention in the last version, the avatar was highly optimized in terms of parameters (a bunch of unnecessary bool variables in “params” were removed. Added “vrcfox Template” for quick rig (avatar configuration). Optimized blend-file by memory slightly. Also renamed a lot of avatar parameters (now they will be clearer than they were before and have been tweaked to fit the overall style). There was a super small redesign of icons and the location of some parameters in the menu.
As it is a pre-release, not everything is finalized yet, it will be up to date soon :D