v1.3 (Pre-release) [v2.0]
Disabling locomotion animation for crouching and prone at will has been added (not tested yet).
Changed project folder names and some file names, and also fixed a small bug with the avatar body (2 double tops were detected), as well as added sharp tops for Boots and tinted it a bit.
The “Authors” menu was removed (who needs excessive advertising? If it's already in the avatar description :D), as well as Thumbnail.png with ThumbnailOverride script for the sake of project memory optimization.