Major improvement of AnimatorWizard and slightly improved locomotion
Added support for adding/removing clothing with special prefixes, will need to improve this algorithm significantly, and add support for automatic creation of OSCmooth (like so now you have to add OSCmooth parameters yourself after the animator setup. It's boring -_-) to FaceTracking (FT) and automatic creation of EyeTracking Layer. Also maybe we need to add a flag that when you use the word “left” in FT in the blendshape name, then it automatically creates for the right side as well (it was there before, but without the flag. I removed it temporarily as there were problems with blendshapes that have the words “left” in the min/max Shape Name).

New animations have been added to Locomotion, which “animate” the avatar movement a bit better :)