bug fix with triangulation
The vertex color looked bad when triangulating the mesh (this must be done in Blender before exporting to get rid of the "A polygon of Mesh 'Body' in Assets/vrcfox model (B&C).fbx is self-intersecting and has been discarded" error). It was a Beauty type, so I replaced it with a Fixed type, which seemed to solve the problem. The triangulation logic itself also changed slightly.